+++ flinkspector-core/src/main/java/org/flinkspector/core/quantify/records/ListQuantifyMatchers.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.flinkspector.core.quantify.list;
+package org.flinkspector.core.quantify.records;
 
 import org.hamcrest.Matcher;
 import org.hamcrest.core.IsNot;
