+++ flinkspector-core/src/main/java/org/flinkspector/core/quantify/records/OnEach.java
@@ -14,48 +14,43 @@
  * limitations under the License.
  */
 
-package org.flinkspector.core.quantify.list;
+package org.flinkspector.core.quantify.records;
 
 import org.hamcrest.Description;
-import org.hamcrest.Factory;
 import org.hamcrest.Matcher;
 
+
 /**
- * Provides a {@link Matcher} that is successful if at most n
- * items in the examined {@link Iterable} is a positive match.
+ * Provides a {@link Matcher} that is successful if each of
+ * the items in the examined {@link Iterable} is a positive match.
  * @param <T>
  */
-public class OnAtMost<T> extends WhileList<T> {
-
-	private final int n;
+public class OnEach<T> extends WhileList<T> {
 
 	/**
-	 * Default constructor
-	 * @param matcher to apply to the {@link Iterable}
-	 *                @param n number of expected matches
+	 * Default Constructor
+	 * @param matcher to apply to {@link Iterable}.
 	 */
-	public OnAtMost(Matcher<T> matcher, int n) {
+	public OnEach(Matcher<T> matcher) {
 		super(matcher);
-		this.n = n;
 	}
 
 	@Override
 	protected Description describeCondition(Description description) {
-		return description.appendText("at most ").appendValue(n);
-	}
-
-	@Override
-	public String prefix() {
-		return "at most " + n +" records";
+		return description.appendText("<all>");
 	}
 
 	@Override
 	public boolean validWhile(int numMatches, int numMismatches) {
-		return numMatches <= n;
+		return numMismatches <= 0;
 	}
 
-	@Factory
-	public static <T> OnAtMost<T> atMost(Matcher<T> matcher, int n) {
-		return new OnAtMost<T>(matcher,n);
+	@Override
+	public String prefix() {
+		return "each record";
+	}
+
+	public static <T> OnEach<T> each(Matcher<T> matcher) {
+		return new OnEach<>(matcher);
 	}
 }
