+++ flinkspector-core/src/main/java/org/flinkspector/core/quantify/records/OnExactly.java
@@ -14,45 +14,57 @@
  * limitations under the License.
  */
 
-package org.flinkspector.core.quantify.list;
+package org.flinkspector.core.quantify.records;
 
 import org.hamcrest.Description;
 import org.hamcrest.Factory;
 import org.hamcrest.Matcher;
 
+
 /**
- * Provides a {@link Matcher}s that is successful if at least one
- * item in the examined {@link Iterable} is a positive match.
+ * Provides a {@link Matcher} that is successful if exactly n
+ * items in the examined {@link Iterable} is a positive match.
+ *
  * @param <T>
  */
-public class OnAny<T> extends UntilList<T> {
+public class OnExactly<T> extends WhileList<T> {
+
+	private final int n;
 
 	/**
-	 * Default Constructor
+	 * Default constructor
+	 *
 	 * @param matcher to apply to the {@link Iterable}
+	 * @param n number of expected positive matches
 	 */
-	public OnAny(Matcher<T> matcher) {
+	public OnExactly(Matcher<T> matcher, int n) {
 		super(matcher);
+		this.n = n;
 	}
 
 	@Override
 	protected Description describeCondition(Description description) {
-		return description.appendText("at least ").appendValue(1);
-	}
-
-	@Override
-	protected boolean validWhen(int matches, int possibleMatches) {
-		return matches == 1;
+		return description.appendText("exactly ").appendValue(n);
 	}
 
 	@Override
 	public String prefix() {
-		return "any record";
+		return "exactly " + n + " records";
+	}
+
+	@Override
+	public boolean validWhile(int matches, int mismatches) {
+		return matches <= n;
+	}
+
+	@Override
+	public boolean validAfter(int numMatches) {
+		return numMatches == n;
 	}
 
 	@Factory
-	public static <T> OnAny<T> any(Matcher<T> matcher) {
-		return new OnAny<>(matcher);
+	public static <T> OnExactly<T> exactly(Matcher<T> matcher, int n) {
+		return new OnExactly<T>(matcher,n);
 	}
-
 }
+
