+++ flinkspector-core/src/main/java/org/flinkspector/core/quantify/records/WhileList.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.flinkspector.core.quantify.list;
+package org.flinkspector.core.quantify.records;
 
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
